m|fix doctor get detail and add fav
