+++ routes/api.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\ClinicController;
 use App\Http\Controllers\DoctorController;
+use App\Http\Controllers\FavController;
 use App\Http\Controllers\JWTValidator;
 use App\Http\Controllers\MobileBannerController;
 use App\Http\Controllers\RatingsController;
@@ -10,6 +11,7 @@ use App\Http\Controllers\SplashscreenMobileController;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\UserpetsController;
 use App\Http\Controllers\WalletController;
+use App\Models\fav;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -114,8 +116,8 @@ Route::post('mobilebanner/togglebanner', [MobileBannerController::class, 'toggle
 
 Route::post('mobilebanner/deletebanner', [MobileBannerController::class, 'deletebanner']);
 
-Route::post('service/fav/add', [ServicefavouriteController::class, 'addfav']);
+Route::post('service/fav/add', [FavController::class, 'addfav']);
 
-Route::post('service/fav/delete', [ServicefavouriteController::class, 'deletefav']);
+Route::post('service/fav/delete', [FavController::class, 'deletefav']);
 
-Route::get('service/fav/userlist', [ServicefavouriteController::class, 'getuserfavlist']);
+Route::get('service/fav/userlist', [FavController::class, 'getuserfavlist']);
